refactor: sidebar menu array
